tutorials/Learning Series 12/requirements.txt
open_clip_torch>=2.24.0
torch>=2.1
torchvision>=0.16
numpy>=1.24
pandas>=2.0
tqdm>=4.66
pyyaml>=6.0
Pillow>=10.0
